Corrected float

+++ Options/SignatureMixedbuildingOptions.cs
@@ -5,9 +5,9 @@ public class SignatureMixedbuildingOptions
     public string Name { get; set; } = null!;
 
     public int Attractiveness { get; set; }
-    public int Wellbeing { get; set; }
-    public int Radius { get; set; }
-    public int CityAttractiveness { get; set; }
+    public float Wellbeing { get; set; }
+    public float Radius { get; set; }
+    public float CityAttractiveness { get; set; }
     public int Noisepollution { get; set; }
 
 }
